+++ app/src/main/res/drawable/ic_playlist_add_black_24dp.xml
@@ -1,0 +1,5 @@
+<vector android:height="30dp" android:tint="#000204"
+    android:viewportHeight="24.0" android:viewportWidth="24.0"
+    android:width="30dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M14,10L2,10v2h12v-2zM14,6L2,6v2h12L14,6zM18,14v-4h-2v4h-4v2h4v4h2v-4h4v-2h-4zM2,16h8v-2L2,14v2z"/>
+</vector>
